+++ source/Data/Intermediate Formats/IF_Colors.cs
@@ -1,3 +1,6 @@
+using System.Text.Json.Serialization;
+using Bungie.Game;
+
 namespace Huragok.Data.IntermediateFormats.Color {
     /// <summary>
     /// <para>Available modes which colors can be represented in.</para>
@@ -15,19 +18,23 @@ namespace Huragok.Data.IntermediateFormats.Color {
         /// <summary>
         /// Value of the red channel, spanning from 0 to 255.
         /// </summary>
-        internal readonly int red;
+        public readonly int red;
         /// <summary>
         /// Value of the green channel, spanning from 0 to 255.
         /// </summary>
-        internal readonly int green;
+        public readonly int green;
         /// <summary>
         /// Value of the blue channel, spanning from 0 to 255.
         /// </summary>
-        internal readonly int blue;
+        public readonly int blue;
+        /// <summary>
+        /// Optional value for the alpha channel, spanning from 0 to 255. If empty, color is assumed to be RGB only.
+        /// </summary>
+        public readonly int? alpha;
         /// <summary>
         /// Hexadecimal representation of this color. #RRGGBB
         /// </summary>
-        internal readonly string hexCode;
+        public readonly string hexCode;
         /// <summary>
         /// The original <see cref="IF_ColorMode"/> of this color, in string format.
         /// </summary>
@@ -39,11 +46,13 @@ namespace Huragok.Data.IntermediateFormats.Color {
         /// <param name="r">Red component. (0-255)</param>
         /// <param name="g">Green component. (0-255)</param>
         /// <param name="b">Blue component. (0-255)</param>
+        /// <param name="a">Optional alpha component. (0-255)</param>
         /// <param name="colormode">The <see cref="IF_ColorMode"/> this color is in.</param>
-        internal IF_Color(int r, int g, int b, IF_ColorMode colormode) {
+        internal IF_Color(int r, int g, int b, int? a = null, IF_ColorMode colormode = IF_ColorMode.PC) {
             this.red = r;
             this.green = g;
             this.blue = b;
+            this.alpha = a;
 
             this.hexCode = $"#{r:X2}{g:X2}{b:X2}";
             this.colorMode = colormode.ToString();